uncommented attribution
Due to the fact that code could be coming from any language, without major work we couldn't safely assume the commenting syntax. Because of this, I figured it would be best to leave it up to the person doing to copying (the programmer, in all likelihood) to comment out the source.

in a future update we could have this as a setting that is off by default.